backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.135.3",
    "google-genai>=1.30.0",
    "groq>=0.31.0",
    "langchain-text-splitters>=0.3.11",
    "pinecone>=8.1.1",
    "pydantic>=2.11.3",
    "pypdf>=6.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "passlib[bcrypt]>=1.7.4",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.48",
    "tiktoken>=0.11.0",
    "uvicorn>=0.42.0",
    "youtube-transcript-api>=1.2.2",
    "python-dotenv>=1.2.2",
]
